✨ [기능 추가] :  독서중인 도서 정보 조회 API
등록된 도서가 있으면 도서 정보를 반환하고, 업승면 도서 등록 페이지 입장 유도

+++ springboot/src/main/java/bookcalendar/server/Domain/Member/Service/MemberServiceImpl.java
@@ -131,7 +131,7 @@ public class MemberServiceImpl implements MemberService {
         // 리프레시 토큰 유효성 검사
 
         // 토큰에서 유저 고유 번호 추출
-        Long memberId = jwtService.extractUserNumberFromToken(refreshRequest.accessToken());
+        Integer memberId = jwtService.extractUserNumberFromToken(refreshRequest.accessToken());
 
         // 전송받은 리프레시 토큰 추출
         String oldRefreshToken = refreshRequest.refreshToken();
